Move parse_integer and parse_unsigned_integer from util::str to style::attr
